refactor alertaccordion to use sds
