+++ src/frontend/src/components/AlertAccordion/style.ts
@@ -1,12 +1,53 @@
 import styled from "@emotion/styled";
-import { Alert, fontHeaderS } from "czifui";
+import { Callout, CommonThemeProps, getIconSizes } from "czifui";
+import ArrowDownIcon from "src/common/icons/IconArrowDownSmall.svg";
+import ArrowUpIcon from "src/common/icons/IconArrowUpSmall.svg";
+import { transparentScrollbars } from "src/common/styles/basicStyle";
 
-export const StyledAlert = styled(Alert)`
-  .MuiAlert-action {
-    align-items: start;
+const smallIcon = (props: CommonThemeProps) => {
+  const iconSizes = getIconSizes(props);
+  return `
+    flex: 0 0 auto;
+    height: ${iconSizes?.s.height}px;
+    width: ${iconSizes?.s.width}px;
+  `;
+};
+
+export const StyledArrowDownIcon = styled(ArrowDownIcon)`
+  ${smallIcon}
+`;
+
+export const StyledArrowUpIcon = styled(ArrowUpIcon)`
+  ${smallIcon}
+`;
+
+export const RowFlexContainer = styled.div`
+  display: flex;
+  height: 100%;
+
+  .MuiCollapse-root {
+    overflow-y: auto;
   }
 `;
 
-export const Title = styled.span`
-  ${fontHeaderS}
+export const ColumnFlexContainer = styled.div`
+  display: flex;
+  flex-direction: column;
+  flex: 1 1 auto;
+  .MuiCollapse-root {
+    ${transparentScrollbars}
+  }
+`;
+
+export const StaticSizeDiv = styled.div`
+  flex: 0 0 auto;
+`;
+
+export const StyledCallout = styled(Callout)`
+  max-height: 250px;
+  width: 100%;
+
+  .MuiAlert-message {
+    width: 100%;
+  }
 `;
