+++ src/frontend/src/views/Upload/components/Metadata/components/ImportFile/components/Alerts/common/style.ts
@@ -6,7 +6,6 @@ import {
   getColors,
   getSpaces,
 } from "czifui";
-import AlertAccordion from "src/components/AlertAccordion";
 
 export const Th = styled.th`
   ${fontHeaderXs}
@@ -32,15 +31,6 @@ export const Td = styled.td`
 
 export const Title = styled.span`
   ${fontHeaderS}
-`;
-
-// To get the table to span the entire width of accordion, we must manipulate
-// MUI internal CSS to force its div that wraps content to use full width.
-// If this gets used regularly, might be better as prop on `AlertAccordion`.
-export const FullWidthAlertAccordion = styled(AlertAccordion)`
-  .MuiAlert-message {
-    width: 100%;
-  }
 `;
 
 export const FullWidthContainer = styled.div`
